requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
streamlit==1.55.0
pypdf==3.16.0
requests==2.32.5
python-multipart==0.0.22
torch==2.2.2
numpy==1.26.4
scikit-learn==1.8.0
python-dotenv==1.0.0
huggingface-hub==0.17.3
